requirements.txt
asgiref==3.7.2
Django==4.2.3
django-multiselectfield==0.1.12
sqlparse==0.4.4
tzdata==2023.3
